Fix bug where user assignment and options popovers were obscured
The z-indexes applied in the previous version were causing card content menu popovers -- including the user assignee popover, and the "..." popover -- to appear underneath subsequent cards in the same column (if any).

This fixes the problem by removing the z-index values from everything except the column drag target regions.

Unfortunately this introduces a minor issue where the area that would be occupied by the "+ Create" button, under the last cardin each column, is not a valid drop target. Accepting this issue for now because it is more minor than the issue of the popover menus being mostly unusable.